services/groq_client.py
"""Groq API client initialization"""
import os
from groq import Groq
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

def get_groq_client():
    """Initialize and return Groq client"""
    api_key = os.getenv("GROQ_API_KEY")
    if not api_key:
        raise ValueError("GROQ_API_KEY not found in environment variables")
    return Groq(api_key=api_key)

def get_api_key():
    """Get the Groq API key"""
    return os.getenv("GROQ_API_KEY")
