requirements.txt
streamlit
requests
groq
pandas
openpyxl
python-dotenv
